tests/wpt/web-platform-tests/fenced-frame/default-enabled-features-allow-self.https.html

<!DOCTYPE html>
<title>Test default permission policy features gating (self)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="resources/utils.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="resources/default-enabled-features-helper.js"></script>

<body>
<script>
promise_test(async(t) => {
  await runDefaultEnabledFeaturesTest(t, true, get_host_info().ORIGIN);
}, 'Same-origin fenced frame loads when feature policies are self');

promise_test(async(t) => {
  await runDefaultEnabledFeaturesTest(t, false, get_host_info().REMOTE_ORIGIN);
}, 'Cross-origin fenced frame does not load when feature policies are self');

promise_test(async(t) => {
  const fencedframe = await attachFencedFrameContext({
        generator_api: "sharedstorage",
        origin: get_host_info().ORIGIN});

  await fencedframe.execute(async () => {
      assert_true(document.featurePolicy.allowsFeature('attribution-reporting'),
          "Attribution reporting should be allowed in the outer fenced frame.");
      const fencedframe_inner = await attachFencedFrameContext({
        generator_api: "sharedstorage",
        origin: get_host_info().REMOTE_ORIGIN});

      await fencedframe_inner.execute(async () => {
        assert_true(
            document.featurePolicy.allowsFeature('attribution-reporting'),
            "Attribution reporting should be allowed in the nested fenced " +
            "frame that's cross origin to its embedder.");
      }, []);
  }, []);
}, 'Fenced frames default feature policies are set to allow *');

promise_test(async(t) => {
  // We do this test the "old fashioned way" because a redirect in a fenced
  // frame remote context will cause it to lose its ability to communicate with
  // the main page (which results in a timeout).
  const page1_key = token();
  const redirect_key = token();

  const urn = await generateURNFromFledge("resources/default-enabled-features-redirect.https.html", [page1_key, redirect_key]);
  const fencedframe = attachFencedFrame(urn);

  // The fenced frame will send its attribution reporting result and then
  // attempt to redirect to a remote origin page.
  const page1_resp = await nextValueFromServer(page1_key);
  assert_equals(page1_resp, "true",
      "Attribution reporting should be enabled on the original page.");

  // There is no API to observe whether the document in the fenced frame loaded
  // or not. Instead, set up a timeout. If the document loads, "FAIL" will be
  // sent to the server. Otherwise "blocked" will be sent after 2 seconds.
  const fencedframe_blocked = new Promise(r => t.step_timeout(r, 1000));
  assert_equals("blocked", await Promise.any([
    nextValueFromServer(redirect_key).then(() => "loaded"),
    fencedframe_blocked.then(() => "blocked")
  ]), "The fenced frame redirect should not be successful.");
}, 'A fenced frame redirected to a page that does not allow feature policies ' +
    'does not navigate');

</script>
</body>
</html>
